Optim: in Watchmen mode, only compute the visible part
Before, the whole canvas was computed and then covered with the background color.
Now, only the visible part of the canvas is computed and the rest of the canvas is filled with the yellow color using a gl.clear instruction, which is way faster.
On a 16/9 screen this gives a 50% performance improvement :)