Shader: fixed artifacts visible with some graphic cards
To compute the noise coming from a certain corner, use absolute coordinates instead of local ones.
This fixes a visual artifact I saw on an old model of graphic card: on coordinates where fract(coords) was extremely close to 0 or 1, hard lines appeared in the gradient noise.
However this is slightly less performant (2% on my measures). This is most likely because I replaced a const expression by a variable expression in the substraction.